Refactor monaco into a component
